+++ CHANGELOG.md
@@ -4,7 +4,20 @@ Change Log
 All notable changes to this project will be documented in this file.
 This project adheres to [Semantic Versioning](http://semver.org/).
 
+
 Unreleased Changes
+******************
+
+12.3.0
+******************
+* Data Model changes to support Holography as per ADR-106 
+  - Configure schema updated for Holography as per ADR-106
+  - Groups key added in pointing configuration
+  - In this release validation added only for mosaic, table, fixed pattern 
+  - Adds HolographyReceptorGroupConfig , TrajectoryConfig and ProjectionConfig to subarray_node/configuration 
+  - Adds MosaicTrajectoryConfig, TableTrajectoryConfig, FixedTrajectoryConfig to support for Table, Mosaic and Fixed pattern
+
+12.2.0
 ******************
 * Data model changes to support ADR-99:
   - Deprecated csp.FSPConfiguration